change usd_balance and btc_balance column type to int

+++ src/main/resources/schema.sql
@@ -2,8 +2,8 @@ CREATE TABLE IF NOT EXISTS account (
     user_id INT AUTO_INCREMENT PRIMARY KEY,
     username VARCHAR(50) NOT NULL UNIQUE,
     email VARCHAR(100) NOT NULL,
-    usd_balance DECIMAL(10, 2) DEFAULT 1000.00,
-    btc_balance DECIMAL(18, 8) DEFAULT 0,
+    usd_balance INT DEFAULT 1000,
+    btc_balance INT DEFAULT 0,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
